requirements.txt
transformers
torch
PyMuPDF
accelerate
groq
dotenv
Werkzeug
Flask-Limiter